implement ledger book module for tracking customer and supplier financial history with custom transactions and balance reporting

+++ app/Models/Product.php
@@ -28,8 +28,8 @@ class Product extends Model
     ];
 
     protected $casts = [
-        'purchase_price' => 'integer',
-        'selling_price' => 'integer',
+        'purchase_price' => 'float',
+        'selling_price' => 'float',
         'tax_percentage' => 'float',
         'quantity' => 'integer',
         'min_stock' => 'integer',
